refactor(cmd/tls): move tls subcommands from cmd to internal/tls
