+++ internal/tls/install/install.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cmd
+package install
 
 import (
 	"github.com/spf13/cobra"
